requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1
requests==2.31.0
openai==1.17.0
